Upg: move secrets to their own page, fix: long secrets pushing buttons outside of viewport, fix: handle errors when saving secrest, fix: column size for secrets
