+++ front/lib/models/workspace.ts
@@ -265,7 +265,7 @@ DustAppSecret.init(
       allowNull: false,
     },
     hash: {
-      type: DataTypes.STRING,
+      type: DataTypes.TEXT,
       allowNull: false,
     },
   },
